backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
reportlab==4.2.2
google-genai==1.25.0
openai==1.54.0

